+++ R/pedSimulate.R
@@ -7,30 +7,26 @@
 #' @author Mohammad Ali Nilforooshan \email{m.a.nilforooshan@gmail.com}
 #'
 #' @description
-#' An R package for simulating a pedigree with genetic merits and phenotypes, starting from a base population (generation 0).
+#' An R package for simulating a pedigree with genetic merits and phenotypes, starting from a base population (generation 0) or an existing pedigree.
 #' The pedigree depth and design can be chosen by the values provided to the arguments of the simulation function.
 #'
 #' @details
-#' Starting from a base population with equal number of males and females,
-#' next generations are simulated for a user-defined number of generations.
-#' The size of the base population is defined by the user.
+#' Starting from a base population with a user-defined size and equal number of males and females,
+#' next generations are simulated for the user-defined littersize and number of generations.
 #' No selection (natural or artificial) and non-random mating is applied to this population.
+#' Alternatively, the simulation can be started from an existing pedigree.
 #' Natural (mortality) and artificial selection are applied to the next generations.
-#' Different selection patterns can be applied to males and females, including
-#' different proportion of selected individuals from selection candidates (after applying mortality) as parents of the next generation,
-#' random or merit-based selection, and generation overlap.
-#' Selected individuals are mated randomly using function \code{\link{simulatePed}},
-#' or mated assortatively/disassortatively using function \code{\link{assortative}}.
-#' In assortative/disassortative mating, males and females are sorted for mating based on the criteria for which they have been selected for
-#' (can be different between males and females).
-#' Further choices, such as litter size, and avoiding fullsib matings and parent-progeny matings are provided.
+#' Different generation overlap, and selection intensities and patterns can be applied to males and females.
+#' Selected individuals are mated randomly or assortatively/disassortatively, 
+#' based on different selection criteria applied to males and females, independently.
 #' Performance and genetic merit of individuals are simulated using the basic rules of quantitative genetics.
 #' The performance (P) of an individual is influenced by genetic (A) and environmental (E) effects.
 #' Thus, P = A + E, and Var(P) = Var(A) + Var(E).
 #' The additive genetic merit (A) of an individual is the average of its parents' additive genetic merits
-#' '(PA = (A\ifelse{html}{\out{<sub>sire</sub>}}{\eqn{_{sire}}} + A\ifelse{html}{\out{<sub>dam</sub>}}{\eqn{_{dam}}})/2)
+#' (PA = (A\ifelse{html}{\out{<sub>sire</sub>}}{\eqn{_{sire}}} + A\ifelse{html}{\out{<sub>dam</sub>}}{\eqn{_{dam}}})/2)
 #' plus the Mendelian Sampling term due to the sampling of alleles passed from the parent to the offspring.
 #' The Mendelian Sampling variance is half of Var(A) in the base population.
 #' Because there is no provided information for environmental effects, the environment effect is
 #' assigned to individuals from a normal distribution of random numbers (E ~ N(0, \strong{I}Var(E))).
+#' The package also provides functions to identify halfsib, fullsib and parent-progeny matings in the pedigree.
 NULL